# core/src/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Swagger - Core microservice documentation",
    "description": ""
  },
...
  "paths": {
    "/store/": {
      "post": {
        "tags": [
          "store"
        ],
        "summary": "Create a new store",
        "description": "",
        "operationId": "createStore",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Store"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/Store"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Store"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Store"
                }
              }
            }
          },
          "405": {
            "description": "Invalid input"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]              
      }
    },
    "/store": {
      "get": {
        "tags": [
          "store"
        ],
        "summary": "Returns all stores",
        "description": "",
        "operationId": "getInventory",
        "responses": {
          "200": {
            "description": "successful operation"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]           
      }
    },
    "/store/{storeId}": {
      "get": {
        "tags": [
          "store"
        ],
        "summary": "Find store by ID",
        "description": "",
        "operationId": "getStoreById",
        "parameters": [
          {
            "name": "storeId",
            "in": "path",
            "description": "ID of store that needs to be fetched",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Store"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Store"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Store not found"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]           
      },
      "put": {
        "tags": [
          "store"
        ],
        "summary": "Update store",
        "description": "",
        "operationId": "updateStore",
        "parameters": [
          {
            "name": "storeId",
            "in": "path",
            "description": "ID of store that needs to be fetched",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Update store",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Store"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/Store"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Store"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "successful operation"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]           
      },
      "delete": {
        "tags": [
          "store"
        ],
        "summary": "Delete store by ID",
        "description": "",
        "operationId": "deleteStore",
        "parameters": [
          {
            "name": "storeId",
            "in": "path",
            "description": "ID of the store that needs to be deleted",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Store not found"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]           
      }
    },
...
    "/event/": {
      "post": {
        "tags": [
          "event"
        ],
        "summary": "Create a new event",
        "description": "",
        "operationId": "createEvent",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Event"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/Event"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Event"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Event"
                }
              }
            }
          },
          "405": {
            "description": "Invalid input"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]          
      }
    },
    "/event": {
      "get": {
        "tags": [
          "event"
        ],
        "summary": "Returns all events",
        "description": "",
        "operationId": "getInventory",
        "responses": {
          "200": {
            "description": "successful operation"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]          
      }
    },
    "/event/{eventId}": {
      "/event/": {
        "post": {
          "tags": [
            "event"
          ],
          "summary": "Create a new event",
          "description": "",
          "operationId": "createEvent",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Event"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Event"
                }
              },
              "application/x-www-form-urlencoded": {
                "schema": {
                  "$ref": "#/components/schemas/Event"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "successful operation",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Event"
                  }
                }
              }
            },
            "405": {
              "description": "Invalid input"
            }
          },
          "security": [
            {
              "api_key": []
            }
          ]            
        }
      },
      "/event": {
        "event": {
          "tags": [
            "event"
          ],
          "summary": "Returns all events",
          "description": "",
          "operationId": "getInventory",
          "responses": {
            "200": {
              "description": "successful operation"
            }
          }
        }
      },  
      "get": {
        "tags": [
          "event"
        ],
        "security": [
          {
            "api_key": []
          }
        ],         
        "summary": "Find event by ID",
        "description": "",
        "operationId": "getEventById",
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "description": "ID of event that needs to be fetched",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Event"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Event"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "event not found"
          },
          "security": [
            {
              "api_key": []
            }
          ]            
        }
      },
      "put": {
        "tags": [
          "event"
        ],
        "security": [
          {
            "api_key": []
          }
        ],         
        "summary": "Update event",
        "description": "",
        "operationId": "updateEvent",
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "description": "ID of event that needs to be fetched",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Update event",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Event"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/Event"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Event"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "successful operation"
          }      
        }
      },
      "delete": {
        "tags": [
          "event"
        ],
        "security": [
          {
            "api_key": []
          }
        ],          
        "summary": "Delete event by ID",
        "description": "",
        "operationId": "deleteStore",
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "description": "ID of the event that needs to be deleted",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "event not found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Store": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "store #1"
          },
          "address": {
            "type": "string",
            "example": "address #1"
          },          
          "latitude": {
            "type": "integer",
            "format": "int32",
            "example": 10
          },
          "longitude": {
            "type": "integer",
            "format": "int32",
            "example": 20
          }
        },
        "xml": {
          "name": "store"
        }
      },
...
      "Event": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "event #1"
          },
          "active": {
            "type": "boolean",
            "example": true
          }          
        },
        "xml": {
          "name": "event"
        }
      }
    },
...
    "securitySchemes": {
      "api_key": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }    
  }
}